Factor out dotpkg resolution
